backend/requirements.txt
alembic==1.14.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
packaging==24.2
PyJWT==2.10.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
requests==2.26.0
google-generativeai==0.8.3
tenacity==8.0.1
SpeechRecognition==3.8.1
pydub==0.25.1
redis==4.5.5
flask-socketio==5.3.6
authlib==1.3.0
sendgrid==6.11.0
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
requests-oauthlib==1.3.1
google-auth==2.23.3
oauthlib==3.2.2
google-api-python-client==2.108.0